backend/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
arxiv==2.1.0
async-timeout==4.0.3
attrs==23.1.0
awslambdaric==2.0.8
beautifulsoup4==4.12.3
blinker==1.6.2
boto3==1.28.9
botocore==1.31.9
Brotli==1.1.0
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.2.0
cleo==2.1.0
click==8.1.7
codetiming==1.4.0
colorama==0.4.6
configparser==6.0.0
crashtest==0.4.1
cssselect==1.2.0
cssselect2==0.7.0
curl_cffi==0.7.0b4
dataclasses-json==0.6.3
decoder==0.5
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
duckduckgo_search==5.3.1
dulwich==0.21.7
fastapi==0.101.0
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
Flask==2.3.2
fonttools==4.51.0
frozenlist==1.4.0
google-ai-generativelanguage==0.6.2
google-api-core==2.11.1
google-api-python-client==2.128.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-generativeai==0.5.2
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib_metadata==7.1.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
keras==2.13.1
keyring==24.3.1
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.56
lxml==5.2.1
lxml_html_clean==0.1.1
Markdown==3.6
markdown2==2.4.13
MarkupSafe==2.1.3
marshmallow==3.20.1
md2pdf==1.0.1
mistune==3.0.2
mlx==0.0.4
more-itertools==10.2.0
motor==3.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.25.1
openai==1.27.0
opencv-python==4.8.0.74
orjson==3.10.3
packaging==23.2
pandas==2.0.3
pexpect==4.9.0
Pillow==10.0.0
pkginfo==1.10.0
platformdirs==4.2.1
playwright==1.43.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.7.1
proto-plus==1.23.0
protobuf==4.23.4
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11 @ file:///private/tmp/pybind11-20231208-64091-6sikag/pybind11-2.11.1
pycipher==0.5.2
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydyf==0.10.0
pyee==11.1.0
pygltflib==1.16.1
pymongo==4.7.3
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
pyparsing==3.1.2
pyphen==0.15.0
pyproject_hooks==1.1.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
QScintilla==2.14.1
rapidfuzz==3.9.0
regex==2024.5.10
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.6.1
sentencepiece==0.1.99
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.17.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.27.0
sympy==1.12
tavily-python==0.3.3
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.6.0
tinycss2==1.3.0
tinysegmenter==0.3
tldextract==5.1.2
tomlkit==0.12.5
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.4
trove-classifiers==2024.4.10
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.26.1
weasyprint==62.1
webencodings==0.5.1
websockets==12.0
Werkzeug==2.3.6
wrapt==1.16.0
xattr==1.1.0
yarl==1.9.2
zipp==3.18.1
zopfli==0.2.3
